requirements.txt
flask
pillow
pandas
google-api-python-client
requests
xlrd
matplotlib
opencv-python